+++ mycodo/scripts/upgrade_commands.sh
@@ -12,7 +12,7 @@ fi
 
 # Required apt packages. This has only been tested with Raspbian for the
 # Raspberry Pi but should work with most Debian-based systems.
-APT_PKGS="fswebcam gawk gcc git libffi-dev libi2c-dev logrotate \
+APT_PKGS="gawk gcc git libffi-dev libi2c-dev logrotate \
           moreutils nginx python-setuptools sqlite3 wget \
           python3 python3-dev python3-smbus python3-pylint-common \
           rng-tools"