+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Add REST API access, API endpoints, and API manual section ([#705](https://github.com/kizniche/mycodo/issues/705))
  - Add button to copy device UUID to clipboard
  - Add ability to set IP, port, and timeout for upgrade internet check
+ - Add new Camera library: opencv
 
 
 ## 7.9.1 (2019.10.26)